Fix drawing for backwards strafing
During backwards strafing it was possible that the values for the
current angle and the good strafing angles differed by more than 2*pi.
In reality the difference of course is much smaller, it just so happened
that the angles did not fit to each other. To fix this, we now make sure
to change the current angle to the equivalent angle, but that is closest
to the good strafing angles.