Avoid using C Standard Library
The Quake 3 Virtual Machine seems to not support using the C Standard
Library. So if we want to support that, it's gonna be easier to avoid
using it as much as possible. So implemented simple functions like fminf
and sign manually. For other functions like sqrtf and acosf don't really
want to do a manual implementation. We still remove the standard header
includes. Now for using these files, a user has to define the functions
sqrtf and acosf with the strafe_helper_includes.h file. That can be
either by simply including the standard header if possible, or using
their own implementation.