More flexible customization of draw functions
Instead of only allowing customization of color, allow full
customization based on what element is to be drawn. That also allows
declaration of those functions that have to be implemented in a header
inside this repository. So that makes it a bit clearer than the rather
vague explanation in the readme from before..