Use atan2 for computing angle between vectors
The formula using atan2 is less error prone on edge cases than the usual
acos formula. The acos formula can easily break down when the norm of
one of the vectors goes to 0, or when the numerical evaluation of the
fraction happens to be slightly outside of [-1.0,1.0] due to floating
point errors. This caused the strafe helper to flicker a lot, especially
when on the ground.

The atan2 formula also already includes the sign of the
angle, which simplifies things, because the sign was added to the acos
formula manually before.