Remove consideration of forward vector
This did not really provide any value and just made things more
confusing (the bad variable name did not help). The only thing this
really affected was the non-centered drawing of the strafe helper. In
that case, the middle of the screen would mean an angle of 0 between
forward and velocity vector, whereas now it means an angle of 0 between
wishdir and velocity vector. While there might be some not super clear
intuition that the former provides, it also makes the strafe helper more
difficult to use with unusual movement inputs (like backwards strafing),
because the drawn angles will be far away from the center. And the
centered drawing of the strafe helper is pretty much always the
preferred variant anyways.