Remove dependency on truncf
The code for making the angles fit together in the same modulo 2pi range
is unnecessary. angle_minimum and angle_maximum are computed with an
acos followed by maybe getting a sign, meaning they are in the range
[0, pi] or [-pi, 0]. angle_current is computed as atan2 which yields
values in the range [-pi, pi]. So the difference between either of them
will never exceed 2pi.